+++ src/Controller/TagController.php
@@ -16,11 +16,11 @@ final class TagController extends AbstractController
 {
 
     #[Route('/new', name: 'app_tag_add', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager,TagRepository $tags): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, TagRepository $tags): Response
     {
         // method pour créer un tag
-        if(!$this->isGranted('ROLE_USER')) return $this->redirectToRoute('app_home');
-        
+        if (!$this->isGranted('ROLE_USER')) return $this->redirectToRoute('app_home');
+
         $tag = new Tag();
         $tag->setUser($this->getUser());
         $form = $this->createForm(TagType::class, $tag);
@@ -40,11 +40,16 @@ final class TagController extends AbstractController
         ]);
     }
 
+    //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     #[Route('/edit/{id}', name: 'app_tag_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Tag $tag, EntityManagerInterface $entityManager): Response
     {
         // method pour modifier un tag
-        if(!$this->isGranted('ROLE_USER')) return $this->redirectToRoute('app_home');
+        if (!$this->isGranted('ROLE_USER')) return $this->redirectToRoute('app_home');
+        if ($tag->getUser() !== $this->getUser()) {
+            return $this->redirectToRoute('app_account', [], Response::HTTP_SEE_OTHER);
+        }
+        // si l'utilisateur n'est pas le propriétaire du tag, il est redirigé vers la page du compte
 
         $form = $this->createForm(TagType::class, $tag);
         $form->handleRequest($request);
@@ -62,13 +67,18 @@ final class TagController extends AbstractController
         ]);
     }
 
+    //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
     #[Route('/{id}', name: 'app_tag_delete', methods: ['POST'])]
     public function delete(Request $request, Tag $tag, EntityManagerInterface $entityManager): Response
     {
         // method pour supprimer un tag
-        if(!$this->isGranted('ROLE_USER')) return $this->redirectToRoute('app_home');
+        if (!$this->isGranted('ROLE_USER')) return $this->redirectToRoute('app_home');
+        if ($tag->getUser() !== $this->getUser()) {
+            return $this->redirectToRoute('app_account', [], Response::HTTP_SEE_OTHER);
+        }
+        // si l'utilisateur n'est pas le propriétaire du tag, il est redirigé vers la page du compte
 
-        if ($this->isCsrfTokenValid('delete'.$tag->getId(), $request->getPayload()->getString('_token'))) {
+        if ($this->isCsrfTokenValid('delete' . $tag->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($tag);
             $entityManager->flush();
             $this->addFlash('success', 'Tag supprimé avec succès');
